Clear unused file and Code Refactoring, Create Makefile
- Clear unused file
- Code Refactoring
- Create Makefile
