--- Makefile_analysis
@@ -1,18 +1,0 @@
-CC = gcc
-CFLAGS = -Wall -Wextra -g
-LINK = -lncurses
-
-TARGET = resources_status_display
-SRCS = 00_main_analysis.c common.c cpu_info_from_log.c mem_info_from_log.c resources_display.c sys_info.c disk_info.c
-OBJS = $(SRCS:.c=.o)
-
-all: $(TARGET)
-
-$(TARGET): $(OBJS)
-	$(CC) $(CFLAGS) -o $@ $^ $(LINK)
-
-%.o : %.c
-	$(CC) $(CFLAGS) -c $< -o $@
-
-clean:
-	rm -f $(OBJS) $(TARGET)
