--- Makefile_collector
@@ -1,17 +1,0 @@
-test_collector.out: 00_main_collector.o cpu_info_to_log.o common.o mem_info_to_log.o
-	gcc -o test_collector.out 00_main_collector.o cpu_info_to_log.o common.o mem_info_to_log.o
-
-00_main.o: common.h cpu_info_to_log.h 00_main_collector.c
-	gcc -c 00_main_collector.c
-
-cpu_info_to_log.o: cpu_info_to_log.c cpu_info_to_log.h
-	gcc -c cpu_info_to_log.c
-
-common.o: common.c common.h
-	gcc -c common.c
-
-mem_info_to_log.o: mem_info_to_log.c mem_info_to_log.h
-	gcc -c mem_info_to_log.c
-
-clean:
-	rm *.o test*.out
