+++ Makefile_mainScreen
@@ -1,0 +1,18 @@
+CC = gcc
+CFLAGS = -Wall -Wextra -Werror
+LINK = -lncurses -lpthread
+
+TARGET = mainScreen
+SRCS = mainScreen.c
+OBJS = $(SRCS:.c=.o)
+
+all: $(TARGET)
+
+$(TARGET): $(OBJS)
+	$(CC) $(CFLAGS) -o $@ $^ $(LINK)
+
+%.o : %.c
+	$(CC) $(CFLAGS) -c $< -o $@
+
+clean:
+	rm -f $(OBJS) $(TARGET)
