--- Makefile_tmpcleanup
@@ -1,52 +1,0 @@
-#makefile made by chatGPT
-#검수 안함
-
-# 컴파일러와 플래그 설정
-CC = gcc
-CFLAGS = -Wall -g -lncurses
-
-# 소스 파일 및 객체 파일
-SRCS = tmp_cleanup.c cleanup.c check_before_cleanup.c
-OBJS = $(SRCS:.c=.o)
-
-# 실행 파일 이름
-TARGET = tmp_cleanup_program
-
-# 자동 의존성 파일 포함
-DEPS = $(SRCS:.c=.d)
-
-# 기본 목표
-all: $(TARGET)
-
-# 실행 파일 만들기
-$(TARGET): $(OBJS)
-	$(CC) $(OBJS) -o $(TARGET) $(CFLAGS)
-
-# 객체 파일 생성 규칙
-%.o: %.c
-	$(CC) -c $< -o $@ $(CFLAGS)
-
-# 의존성 파일 생성 (헤더 파일 변경 시 자동으로 반영)
-%.d: %.c
-	$(CC) -M $(CFLAGS) $< > $@
-
-# 의존성 파일 포함
--include $(DEPS)
-
-# 청소 규칙
-clean:
-	rm -f $(OBJS) $(DEPS) $(TARGET)
-
-# 실행 파일 실행
-run: $(TARGET)
-	./$(TARGET)
-
-# 디버그 모드에서 빌드
-debug: CFLAGS += -DDEBUG
-debug: $(TARGET)
-
-# 테스트 실행 (선택 사항, 예시로 추가)
-test: $(TARGET)
-	./$(TARGET)
-
-.PHONY: all clean run debug test
