--- README.md
@@ -1,56 +1,0 @@
-## 시스템 프로그래밍 7팀 - 서버 관리 프로그램
-* 시스템프로그래밍 7팀 프로젝트의 최종 결과물입니다.
-* The semester of course: 2024-Fall / ELEC0462-003
-* Professor: Jeong, Chang-Su
-* 개발 기간: 2024. 11. 10. - 2024. 12. 00.
-* 개발 & 테스트 환경: Ubuntu 24.04, Rocky Linux 9.4
-
-### 팀원 구성
-* 권태허 ()
-* 문예강 ()
-* 서형철 (컴퓨터학부 심화컴퓨터공학전공)
-* 우상욱 ()
-* 정혜민 ()
-[Role & Pesponsibilities Image 첨부 - Slide #04]
----------------------------------
-
-### 라인 수 (Powered by LogMetrics)
-- 약 xxx Lines. (Excluding Comments.)
-
-### 세부 기능 적용 여부
-* Thread:
-* File I/O:
-* Signal Handling:
-* Pipe: 
-* Multi-Threading:
-* Other: ???
-
-### FlowChart
-- 각 기능별로 FlowChart 나누어서 그릴지? 
-- 그냥 FlowChart 없이 Block Diagram 한장으로 퉁칠지?
-
----------------------------------
-
-### 프로그램 소개
-* 편리하게 서버를 점검할 수 있도록, 몇몇 기능들을 통합한 프로그램을 제작하였습니다.
-* 프로그램이 제공하는 기능은 아래와 같습니다.
-+ 중요 File Permission 적절성 점검 & 적절 Permission 으로 수정
-+ (Connection ?) Log check & logrotate
-+ Server Resources(CPU, RAM, Storage) Status 점검 & 이상값 로그 기록
-+ User Password 변경 주기 점검
-+ Temporary File 정리
-
-## 프로그램 사용 방법
-- 실행 이미지 첨부하기 (Default)
-- 중요 File: File List 안내 및 Permission 종류 기록
-- Log check: 어떤 Log 점검하는지 기록.
-- Resources: Refresh 간격, Usage 수집방법, 로그기록 방법(Form), 수집 Factor 등 안내.
-- User Password: 변경 주기 기록
-- Temp. File: 어떤 File 정리하는지 기록.
-
-### Additionally, and so on?
-
----------------------------------
-
-### Last Changed: 2024. 11. 10. (Sun.)
-- First Created: 2024. 11. 10. (Sun.)
