--- h_for_tmp_cleanup.h
@@ -1,20 +1,0 @@
-#ifndef H_FOR_TMPCLEANUP.H
-
-#define H_FOR_TMPCLEANUP.H
-
-#include <stdio.h>
-#include <stdlib.h>
-#include <string.h>
-#include <sys/types.h>
-#include <sys/stat.h>
-#include <unistd.h>
-#include <dirent.h>
-#include <fcntl.h>
-#include <time.h>
-#include <errno.h>
-#include <pwd.h>
-#include <grp.h>
-#include <sys/file.h>
-#include <curses.h>
-
-#endif
