--- mem_info_from_log.h
@@ -1,10 +1,0 @@
-#ifndef MEM_INFO_FROM_LOG_H
-
-#include "common.h"
-#include "mem_info_struct.h"
-
-#define MEM_INFO_FROM_LOG_H
-
-MEM_Info get_Mem_Information(int);
-
-#endif
